[RENAME] 패키지명 오타 수정
